Retain proper indentation of multiline docstrings
This change fixes an issue where multiline docstrings would have their
body contents indented a level beyond where they should have been,
causing potential rendering issues with Sphinx and the REPL's builtin
`help` functionality.

Rather than blindly joining our deprecation details to the docstring, we
now look to see if it's a multiline docstring or not, and if it is, we
more carefully append to it. In order to preserve the intended
indentation, we split the docstring into its summary line and the rest
of the contents, use `textwrap.dedent` on the contents, and then join
the summary and contents back up along with the details of the
deprecation message so they'll all be on the same level.

This includes a change to sample.py with a multiline docstring example,
which the docs will render.

Fixes #10