Use the newline index to split on our own
While this introduces no performance benefit—the benchmarks I ran were
almost exactly the same between this change and the initial
implementation—since we already know where we want to split based on the
first newline, just use that for getting us the contents. There's no
need to do a splitlines and then immediately rejoin a majority of those
strings and then do yet another join on top of that.