Reference the sample module more obviously (#26)
Previously, the sample module was hidden in the middle of a paragraph,
though if you're reading the section it'll make sense. Since this is a
single page and the TOC isn't as helpful there with a separate page,
include a reference to the sample module at the very top to catch
people's attention more easily.